refactor: Modify existing files, add Admin & Event controllers

+++ src/main/java/com/eventsphere/services/servicesImpl/EventServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -51,7 +52,34 @@ public class EventServiceImpl implements EventService {
     }
 
     @Override
-    public void deleteEvent(Long eventId) {
-        eventRepository.deleteById(eventId);
+    public Event updateEvent(Long eventId, Event updatedEvent) {
+        System.out.println(eventRepository.findById(eventId));
+        Event existingEvent = eventRepository.findById(eventId)
+                .orElseThrow(() -> new RuntimeException("Event not found"));
+
+        existingEvent.setEventName(updatedEvent.getEventName());
+        existingEvent.setDescription(updatedEvent.getDescription());
+        existingEvent.setEventDateTime(updatedEvent.getEventDateTime());
+        existingEvent.setVenue(updatedEvent.getVenue());
+        existingEvent.setCreatedBy(updatedEvent.getCreatedBy());
+        existingEvent.setAttendees(updatedEvent.getAttendees());
+        existingEvent.setCrewMembers(updatedEvent.getCrewMembers());
+        existingEvent.setSpeakers(updatedEvent.getSpeakers());
+        existingEvent.setUpdatedAt(updatedEvent.getUpdatedAt());
+
+        return eventRepository.save(existingEvent);
+    }
+
+    @Override
+    public Event deleteEvent(Long eventId) {
+        //eventRepository.findById(eventId).orElseThrow(()-> new RuntimeException("event not found"));
+        if(eventRepository.existsById(eventId)){
+            Event event = eventRepository.findById(eventId).get();
+            eventRepository.deleteById(eventId);
+            return event;
+
+        }
+        return null;
+        //return "Event Cannot be deleted,event not found";
     }
 }
